A Whole Bunch Of Stuff
I've been working while on vacation for the past week or so, and so a large amount of work has been backlogged and not committed. Most of it is broken right now so this version may not be in any sort of working order. But in order to clear up my table and fix the problems one at a time with commits for each fix, I am committing all of the changes now. This may be stupid, but I really think I'm the only one who will ever look at this repo. Anyway, a few of the changes are working towards divorcing the Audio backend with SoLoud and miniaudio from the main tree, adding more tilemap support, including the new Grid class for tilemap asset loading, and more.