So Much is Broken This Commit...
I know I shouldn't but I'm committing a version of the raycollisions which is genuinely broken and does not work in any sense. I can't figure out an obvious fix so I'm feeling the need for more diagnostic tools, such as some visual debugging. So I'm quickly committing a broken version, so I can implemement some visual debugging code in the next commit.