A Change Of Focus
I think that the scope of Glewy has grown too large. It was supposed to be about being able to render graphics with opengl much simpler. But for some reason, I kept adding functionality which seems to be more along the lines of writing a terrible game engine. Audio, Scenes, Content Loading, Entities, Components, and more, all of which are unnecessary for just rendering and writing shaders. So from this point on, I'm looking to fix that issue. I want to move functionality that isn't needed, such as Audio, Content, and Scene Management, into their own repositories, like has been done with moving Utilities and Structures to substd. And then focus on making a much simpler yet powerful rendering pipeline for my own general purpose use.